[runtime] Make mini-llvm use only mono_aot_can_dedup (mono/mono#10084)
LLVM was using the helper for the abandoned prototype linkonce work that
much predated when I picked up the project.

This is a problem because it can lead to aot-compiler thinking a method
is dedupable, while llvm thinking it's not. When this happens, the
symbol does not have the Assembly baked into it's name to ensure
uniqueness.

Commit migrated from https://github.com/mono/mono/commit/7edc1024db35ea6d60e6f8e8815f68fa243ef161