Remove `Partial` from `EnablePartialNgenOptimization` (dotnet/corefx#36962)
Ngen optimizations are controlled with two properties - one is EnableNgen, the other is ApplyNgen. We were already using the ApplyNgen property with value of `full` (as of dotnet/corefx#36450), so this was already doing full Ngen, but we still used EnablePartialNgen to control the EnableNgen property. It's a roundabout way to EnableNgen, but does the same thing. This pull request removes the indirection.

Commit migrated from https://github.com/dotnet/corefx/commit/7066a0cac2cadd21ca144cc78d1b1fab9b1f9cd7