Fix MultipleCallsToGetSpan test with pinning (dotnet/corefx#37373)
The test is getting two pointers to an array and validating that they're the same pointer.  But if the array is moved by the GC between the two calls, the test fails.  The fix is to pin the array.

Commit migrated from https://github.com/dotnet/corefx/commit/1a2401cd4c3f56c022173d2ac1e0c34799d387e3