Annotate System.Runtime.CompilerServices.VisualC for nullable reference types (dotnet/corefx#41858)
This assembly really doesn't matter for nullability, as it's only meant to be used with C++/CLI and only by infrastructure.  There's also no annotations necessary on any of the contained attributes, as they are either parameterless or take a parameter that shouldn't be null.  Even so, we may as well mark the assembly as being annotated, so that its status is appropriately reflected by tools.

Commit migrated from https://github.com/dotnet/corefx/commit/57e30310253d9c720722f4363fb75e7df93efa67